Fix offset bug when page refreshes
Use a coroutine job to cancel and restart finding the
current index when the article count changes.
